core/src/test/java/org/neo4j/graphalgo/core/SignatureToolTest.java
/*
 * Copyright (c) 2017-2020 "Neo4j,"
 * Neo4j Sweden AB [http://neo4j.com]
 *
 * This file is part of Neo4j.
 *
 * Neo4j is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
package org.neo4j.graphalgo.core;

import org.junit.jupiter.api.Test;

import java.io.IOException;
import java.security.InvalidKeyException;
import java.security.NoSuchAlgorithmException;
import java.security.SignatureException;
import java.security.spec.InvalidKeySpecException;

import static org.junit.jupiter.api.Assertions.assertTrue;

class SignatureToolTest {

    @Test
    void signAndVerify() throws
        IOException,
        NoSuchAlgorithmException,
        InvalidKeySpecException,
        InvalidKeyException,
        SignatureException {

        String message = "MySuperSecretMessage";

        SignatureTool signatureTool = new SignatureTool();
        var signedMessage = signatureTool.sign(message);
        System.out.println(signedMessage);
        assertTrue(signatureTool.verify(signedMessage));
    }

}